bring all ku data together
